Creates dependecy free template
The current swift4 template requires Alamofire to use many of the
features of the app. Many teams have a "dependency minimization"
mandate and avoid deps such as this one. Alamofire in particular is
unfamiliar to many teams in how it handles certain error cases as well.

If nothing else, this version can eventually show how to separate out
the alamofire parts of the project from the non-alamofire parts.